"Remove TestNG dependency and add executablePath config"
Removed the TestNG dependency from pom.xml file as it is no longer used in the project. Added the 'executablePath' configuration to the Configuration.java and PlaywrightWrapper.java files for more flexibility in specifying browser's executable path. This is especially useful when running the scripts in different environments where the browser's executable path may vary.